LayoutTests/platform/ios-simulator/editing/spelling/spelling-attribute-change-expected.txt

FAIL: Timed out waiting for notifyDone to be called
Tests if the spellchecker behavior change after the spellcheck attribute changed by the script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.spellcheck is true
Hello,zz 
